requirements.txt
numpy==1.21.5
pycryptodome==3.15.0
scikit_image==0.19.2
